+++ docs/changes.rst
@@ -64,6 +64,11 @@ Released: not yet
   - If the private key file is protected with a password, the password prompt
     now states the path name of the private key file in the prompt message.
 
+  - Add optional initialization parameter `url` to pywbem_mock
+    FakedWBEMConnection class. This allows a different URL than the default
+    http://FakedWBEMConnection:5988. With this, tests can be executed with
+    multiple simultaneous mock environments pywbem. (See issue #2711)
+
 * Test: Added support for validating the structure of user-defined properties in
   the easy-server server and vault files. As part of that, increased the minimum
   version of the 'pytest-easy-server' package to 0.8.0. (issue #2660)